Incorporar y eliminar a cadena efectivizados, venta empezada

+++ Caso_de_estudio/src/Estructuras/CadenaDeSupermercados.java
@@ -1,6 +1,7 @@
 package Estructuras;
 
 import Exceptions.SucursalNotFound;
+import org.apache.commons.lang3.StringUtils;
 
 public class CadenaDeSupermercados {
 
@@ -37,8 +38,14 @@ public class CadenaDeSupermercados {
         // Inserto en todas las sucursales de la listaSucursales
         while (actual != null) {
             Sucursal suc = actual.getDato();
-            suc.insertarProducto(prod);
-            actual = actual.getSiguiente();
+            if(suc.getArbolProductos().buscar(prod.getEtiqueta()) == null){
+                suc.insertarProducto(prod);
+                actual = actual.getSiguiente();
+            }
+            else {
+                actual = actual.getSiguiente();
+            }
+            
         }
     }
 
@@ -56,14 +63,26 @@ public class CadenaDeSupermercados {
     }
 
     public void eliminarProductoEnCadena(Comparable Etiqueta) {
-
         Nodo<Sucursal> actual = this.listaSucursales.getPrimero();
         // Inserto en todas las sucursales de la listaSucursales
         while (actual != null) {
             Sucursal suc = actual.getDato();
-            suc.eliminarProducto(Etiqueta);
-            actual = actual.getSiguiente();
+            if(suc.getArbolProductos().buscar(Etiqueta) != null){
+                suc.eliminarProducto(Etiqueta);
+                actual = actual.getSiguiente();
+            }
+            else {
+                actual = actual.getSiguiente();
+            }
+            
         }
+//        Nodo<Sucursal> actual = this.listaSucursales.getPrimero();
+//        
+//        while (actual != null) {
+//            Sucursal suc = actual.getDato();
+//            suc.eliminarProducto(Etiqueta);
+//            actual = actual.getSiguiente();
+//        }
 
     }
 
@@ -75,6 +94,36 @@ public class CadenaDeSupermercados {
             aux.getDato().eliminarProducto(codigo);
         } catch (NullPointerException e) {
             throw new SucursalNotFound();
+
+        }
+
+    }
+
+    public Lista<Integer> VenderProductoEnSucursal(Comparable codigo, Integer cantidad, String suc) throws SucursalNotFound {
+        Lista output = new Lista();
+        //Busco la sucursal para hacer la venta, si no la encuentro devuelvo la excepcion
+        Nodo<Sucursal> aux = this.listaSucursales.buscar(suc.toUpperCase());
+        try {
+            aux.getDato();
+        } catch (NullPointerException e) {
+            throw new SucursalNotFound();
+        }
+        //Si la encuentro me fijo si se puede vender, si se puede, hago la venta
+        if (aux.getDato().sePuedeVender(codigo, cantidad)) {
+            aux.getDato().vender(codigo, cantidad);
+            return null;
+        } else {
+            //Si no se puede vender recorro la lista de sucursales y genero el output
+            Nodo<Sucursal> actual = this.listaSucursales.getPrimero();
+            while (actual != null) {
+                Sucursal sucActual = actual.getDato();
+                if (sucActual.sePuedeVender(codigo, cantidad)) {
+                    TElementoAB<Producto> elem = sucActual.getArbolProductos().buscar(codigo);
+                    Nodo nuevoNodo = new Nodo(sucActual.getNombre(),elem.getDatos().getStock());
+                    output.insertar(nuevoNodo);
+                }
+            }
+            return output;
 
         }
